modification

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: jreyes-s <jreyes-s@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/12 16:23:23 by jreyes-s          #+#    #+#             */
-/*   Updated: 2026/01/15 13:53:18 by jreyes-s         ###   ########.fr       */
+/*   Updated: 2026/01/20 16:08:55 by jreyes-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	size_t len;
+	int	len;
 
 	len = 0;
 	while (s[len] != '\0')
